Revert "build(deps): upgrade datafusion, arrow, and other deps (#504)"
This reverts commit 7a275c68a3cc053f93b42d8eba2344ae575b76cf.

+++ crates/core/src/storage/file_metadata.rs
@@ -23,7 +23,7 @@ pub struct FileMetadata {
     pub name: String,
 
     /// Size in bytes on storage
-    pub size: u64,
+    pub size: usize,
 
     /// Size in bytes in memory
     pub byte_size: i64,
@@ -36,7 +36,7 @@ pub struct FileMetadata {
 }
 
 impl FileMetadata {
-    pub fn new(name: impl Into<String>, size: u64) -> Self {
+    pub fn new(name: impl Into<String>, size: usize) -> Self {
         Self {
             name: name.into(),
             size,
